hrms-app/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose
passlib
bcrypt
python-multipart
alembic
pymysql
slowapi
python-jose[cryptography]